+++ R/app-ui.R
@@ -132,7 +132,16 @@ app_ui <- function(request) {
                       "Validate"
                     )
                   )
+                ),
+                hr(),
+                # Add button to reset the form
+                shinyjs::disabled(
+                  actionButton(
+                    "reset_btn_validate",
+                    "Reset"
+                  )
                 )
+
               ),
 
               # Main panel